Load 3.3.562 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@1195 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/interpret.c
@@ -8830,7 +8830,7 @@ again:
                 ERRORF(("Illegal 'reserve' value for catch(): got %ld, "
                         "required are at least %ld.\n"
                        , sp->u.number
-                       , MASTER_RESERVED_COST * 2
+                       , (long)MASTER_RESERVED_COST * 2L
                        ));
             reserve_cost = sp->u.number;
             sp--;
